requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.12
dynaconf==3.1.7
idna==3.3
oauthlib==3.2.0
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
urllib3==1.26.9

django-ckeditor~=6.3.2